Refine text underline mixin for cross browser goodness
This version is modified from current Medium styles in a different direction
than the last; it renders more consistently and better in isolation (esp on
Firefox/Mac).

I do lose automatic underline resizing (gradient is not sized in em), although
Medium probably eschews that for a (similar? cross browser prettiness?) reason.

It looks like it will also lend itself to decomposition better than the previous
version too, but I'll cross that bridge when I come to it.